move: Layout 파일구조 수정
