+++ src/app/(layout)/(local)/local/[region]/page.tsx
