+++ src/app/(layout)/(local)/local/details/[id]/page.tsx
