+++ src/app/(layout)/search-page/layout.tsx
