+++ src/app/layout.tsx
@@ -28,13 +28,7 @@ function RootLayout({ children }: Readonly<{ children: React.ReactNode }>) {
       <body className={`${pretendard.variable} font-pretendard`}>
         <QueryProvider>
           <div className="max-w-[375px] mx-auto">
-            <Suspense>
-              <HeaderLayout />
-            </Suspense>
-            <div className="pt-[50px] pb-[84px] bg-gray0">
-              <ToastProvider>{children}</ToastProvider>
-            </div>
-            <Footer />
+            <ToastProvider>{children}</ToastProvider>
           </div>
         </QueryProvider>
       </body>
