+++ src/components/MainPage/Main/Main.tsx
@@ -5,7 +5,7 @@ import LocalSection from './section/LocalSection';
 
 function Main() {
   return (
-    <main className="flex flex-col w-full">
+    <main className="flex flex-col w-full bg-gray0">
       <LocalItemList />
       <WebFeedSection />
       <LocalSection />
